Refactor SharedClientTheme into its own module
